FEAT : Course 생성 API 구현 (#32)
* feat : create repositories

* rename : erase MemberStat

* refactor : refactor packaging

* fix : modify README.md

* fix : add dependency for swagger

* fix : add category name at Category

* fix : add takenNumber for calculate number of taken students

* fix : add query function that find CourseList by id

* fix : add ErrorStatus about take, member, course

* fix : add function findByDepartmentId

* fix : add function check member is not SWConvergence

* fix : add category field and some functions

* feat : create CourseException

* feat : create TakeException

* feat : create CourseFilter to check course classification by member

* feat : create ResponseDTO about Take

* feat : create RequestDTO about take

* feat : create TakeConverter to convert DTO and entity

* feat : create TakeService Perform Take-related business logic

* rename : move package

* feat : create a TakeController that processes and make responses to Take-related requests

* feat : create a DashboardController that processes and make responses at Dashboard

* refactor : Change class name MemberStatsDTO to DashboardDTO

* add : gitignore에 yml 추가

* add : dependency에 h2추가

* fix : Category에 문자열 띄어쓰기 수정

* add : course에 수강인원 변동 함수 추가

* add : 사용자가 핵심교양 해당영역을 이미 수강했는지 검사하는 로직 추가

* add : 수업 이름과 학수번호로 미수강 과목 검색하는 쿼리 추가

* fix : 테이블 이름 수정

* fix : 테이블 이름 수정에 따른 레포지토리명 수정

* add : 검색 API 추가

* add : 검색 API 추가에 따른 request 추가

* add : 검색 API 추가에 따른 로직 추가

* fix : fix import

* refactor : rename CurriculumMap to CurriculumTable

* add : add print log of error logic

* fix : fix update logic

* fix : add print log logic

* fix : 패스 과목은 성적 반영X

* fix : request dto 수정

* fix : 중복된 수강 예외 처리

* feat : create TakeService interface

* refactor : Take 패키지 구조 변경

* feat : Course 생성 API 구현

* feat : Course 생성 API 구현

* fix : gitignore에 courses_cse.json 추가
