+++ .gitignore
@@ -40,3 +40,5 @@ out/
 
 ### property file ###
 ./src/main/resources/appliction.yml
+./src/main/resources/courses_cse.json
+
