+++ src/main/java/goorming/iCurriculum/course/exception/CourseException.java
@@ -1,4 +1,4 @@
-package goorming.iCurriculum.course;
+package goorming.iCurriculum.course.exception;
 
 import goorming.iCurriculum.common.code.BaseErrorCode;
 import goorming.iCurriculum.common.exception.GeneralException;
