+++ src/main/java/goorming/iCurriculum/course/repository/CourseRepository.java
@@ -1,5 +1,6 @@
-package goorming.iCurriculum.course;
+package goorming.iCurriculum.course.repository;
 
+import goorming.iCurriculum.course.entity.Course;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
