+++ src/main/java/goorming/iCurriculum/department/DepartmentRepository.java
@@ -3,5 +3,6 @@ package goorming.iCurriculum.department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DepartmentRepository extends JpaRepository<Department,Long> {
-    public Department findDepartmentByName(String department_name);
+    Department findDepartmentByName(String department_name);
+    Department findDepartmentByCode(String code);
 }
